requirements.txt
langchain~=0.3.26
psycopg2-binary~=2.9.10
langchain-core~=0.3.68
langchain-community~=0.3.27
langchain-openai~=0.3.27
python-dotenv~=1.1.1